Clean: 不要な共通コンポーネントの削除。コンポーネントの整理
